# src/main/java/com/sinolec/challenge/model/ProjectGenerationTask.java
package com.sinolec.challenge.model;

import com.fasterxml.jackson.annotation.JsonIgnore;

import javax.persistence.Entity;

@Entity
public class ProjectGenerationTask extends GeneralTask {

    @JsonIgnore
    private String storageLocation;

    public String getStorageLocation() {
        return storageLocation;
    }

    public void setStorageLocation(String storageLocation) {
        this.storageLocation = storageLocation;
    }
}